+++ tests/sdk/test_list.py
@@ -14,11 +14,6 @@ from ..utils import mock_requests
     [
         "dataset",
         "algo",
-        "predicttuple",
-        "testtuple",
-        "traintuple",
-        "aggregatetuple",
-        "composite_traintuple",
         "compute_plan",
         "data_sample",
         "model",
@@ -38,15 +33,32 @@ def test_list_asset(asset_type, client, mocker):
 
 
 @pytest.mark.parametrize(
+    "asset_type",
+    [
+        "predicttuple",
+        "testtuple",
+        "traintuple",
+        "aggregatetuple",
+        "composite_traintuple",
+    ],
+)
+def test_list_task(asset_type, client, mocker):
+    item = getattr(datastore, asset_type.upper())
+
+    mocked_response = make_paginated_response([item])
+    m = mock_requests(mocker, "get", response=mocked_response)
+
+    response = client.list_task()
+
+    assert response == [models.SCHEMA_TO_MODEL[schemas.Type.Task](**item)]
+    m.assert_called()
+
+
+@pytest.mark.parametrize(
     "asset_type,filters",
     [
         ("dataset", {"permissions": ["foo", "bar"]}),
         ("algo", {"owner": ["foo", "bar"]}),
-        ("predicttuple", {"rank": [1, 3]}),
-        ("testtuple", {"rank": [1, 3]}),
-        ("traintuple", {"key": ["foo", "bar"]}),
-        ("aggregatetuple", {"worker": ["foo", "bar"]}),
-        ("composite_traintuple", {"owner": ["foo", "bar"]}),
         ("compute_plan", {"name": "foo"}),
         ("compute_plan", {"status": [models.ComputePlanStatus.done.value]}),
         ("model", {"owner": ["MyOrg1MSP"]}),
@@ -65,6 +77,28 @@ def test_list_asset_with_filters(asset_type, filters, client, mocker):
     m.assert_called()
 
 
+@pytest.mark.parametrize(
+    "asset_type,filters",
+    [
+        ("predicttuple", {"rank": [1, 3]}),
+        ("testtuple", {"rank": [1, 3]}),
+        ("traintuple", {"key": ["foo", "bar"]}),
+        ("aggregatetuple", {"worker": ["foo", "bar"]}),
+        ("composite_traintuple", {"owner": ["foo", "bar"]}),
+    ],
+)
+def test_list_task_with_filters(asset_type, filters, client, mocker):
+    item = getattr(datastore, asset_type.upper())
+
+    mocked_response = make_paginated_response([item])
+    m = mock_requests(mocker, "get", response=mocked_response)
+
+    response = client.list_task(filters)
+
+    assert response == [models.SCHEMA_TO_MODEL[schemas.Type.Task](**item)]
+    m.assert_called()
+
+
 def test_list_asset_with_filters_failure(client, mocker):
     items = [datastore.ALGO]
     m = mock_requests(mocker, "get", response=items)
@@ -77,6 +111,19 @@ def test_list_asset_with_filters_failure(client, mocker):
     assert str(exc_info.value).startswith("Cannot load filters")
 
 
+def test_list_compute_plan_with_ordering(client, mocker):
+    item = datastore.COMPUTE_PLAN
+
+    mocked_response = make_paginated_response([item])
+    m = mock_requests(mocker, "get", response=mocked_response)
+
+    order_by = "start_date"
+    response = client.list_compute_plan(order_by=order_by)
+
+    assert response == [models.SCHEMA_TO_MODEL[schemas.Type.ComputePlan](**item)]
+    m.assert_called()
+
+
 @pytest.mark.parametrize(
     "asset_type",
     [
@@ -85,20 +132,18 @@ def test_list_asset_with_filters_failure(client, mocker):
         "traintuple",
         "aggregatetuple",
         "composite_traintuple",
-        "compute_plan",
     ],
 )
-def test_list_asset_with_ordering(asset_type, client, mocker):
+def test_list_task_with_ordering(asset_type, client, mocker):
     item = getattr(datastore, asset_type.upper())
-    method = getattr(client, f"list_{asset_type}")
 
     mocked_response = make_paginated_response([item])
     m = mock_requests(mocker, "get", response=mocked_response)
 
     order_by = "start_date"
-    response = method(order_by=order_by)
+    response = client.list_task(order_by=order_by)
 
-    assert response == [models.SCHEMA_TO_MODEL[schemas.Type(asset_type)](**item)]
+    assert response == [models.SCHEMA_TO_MODEL[schemas.Type.Task](**item)]
     m.assert_called()